添加微信验证模块
